% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crop_cost.R
\name{calc_crop_cost}
\alias{calc_crop_cost}
\title{calc_crop_cost}
\usage{
calc_crop_cost(crop_rain, mean_rain, risk_parameter, total_harvest, price)
}
\arguments{
\item{crop_rain}{data set with rain by year (inches)}

\item{mean_rain}{the California average rainfall amount (inches)}

\item{risk_parameter}{a paramter that esimates the perentage of crop loss based on deviation from average temperature}

\item{total_harvest}{total expected harvest (lbs)}

\item{price}{price per poound of harvest}
}
\value{
Estimated crop loss or gain, and the monetary value of that loss or gain
}
\description{
This function will find the anomaly in rainfall, and then calculate the estimate loss or gain in crop harvest and the corresponding value of that crop
}
