% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extreme_anom.R
\name{calc_extreme_anom}
\alias{calc_extreme_anom}
\title{calc_extreme_anom}
\usage{
calc_extreme_anom(temp, rain)
}
\arguments{
\item{rain}{}
}
\value{
matrix of results
}
\description{
computes max and min anomalies and what year these anomalies occured for temperature and rain dataframes
}
