% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extreme_rain_anom.R
\name{calc_extreme_rain_anom}
\alias{calc_extreme_rain_anom}
\title{calc_extreme_rain_anom}
\usage{
calc_extreme_rain_anom(rain)
}
\arguments{
\item{rain}{}
}
\value{
list of max and min anomalies and their years
}
\description{
computes max and min anomalies and what year these anomalies occured for precipitation
}
