% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extreme_temp_anom.R
\name{calc_extreme_temp_anom}
\alias{calc_extreme_temp_anom}
\title{calc_extreme_temp_anom}
\usage{
calc_extreme_temp_anom(temp)
}
\arguments{
\item{temp}{}
}
\description{
computes max and min anomalies and what year these anomalies occured for temperature
}
